requirements.txt
Django>=3.0,<4.0
djangorestframework
django-filter
requests
psycopg2-binary>=2.8